secondo commit

+++ richiesteConsapBE/src/main/java/com/proggettazione/richiesteConsapBE/repository/RichiestaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RichiestaRepository extends JpaRepository<Richiesta, Long> {
+public interface RichiestaRepository extends JpaRepository<Richiesta, Integer> {
 }
